+++ core/js-native/src/main/scala/JwtCorePlatform.scala
@@ -1,0 +1,3 @@
+package pdi.jwt
+
+trait JwtCorePlatform[H, C] extends JwtCoreFunctions { self: JwtCore[H, C] => }
