+++ core/js-native/src/main/scala/JwtUtilsPlatform.scala
@@ -1,0 +1,3 @@
+package pdi.jwt
+
+trait JwtUtilsPlatform
