--- core/native/src/test/scala/JwtUtilsPlatformSpec.scala
@@ -1,3 +1,0 @@
-package pdi.jwt
-
-trait JwtUtilsPlatformSpec
