--- json/common/native/src/main/scala/JwtJsonCommonPlatform.scala
@@ -1,3 +1,0 @@
-package pdi.jwt
-
-trait JwtJsonCommonPlatform[J, H, C] extends JwtCore[H, C]
